move expect_used, filter_next, get_unwrap, ok_expect and unwrap_used to their own modules
